Move message handlers to lib/ level
