--- lib/memach/message_handlers/base_handler.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Memach
-  module MessageHandlers
-    class BaseHandler
-      include Import['telegram.bot']
-
-      def self.handle(_message)
-        raise NotImplementedError
-      end
-    end
-  end
-end
